+++ libs/features/spectrum2-styles/build.mjs
@@ -1,9 +1,10 @@
+/* eslint-disable no-nested-ternary */
 import fs from 'fs';
 import path from 'path';
 import postcss from 'postcss';
 
 const spectrumCSSPath = path.resolve('node_modules/@spectrum-css/tokens/dist/index.css');
-const miloCSSPath = path.resolve('../../deps/spectrum2-styles.css');
+const miloCSSPath = path.resolve('../../styles/styles.css');
 
 const logFileSize = (filePath, status) => {
   const fileSizeInBytes = fs.statSync(filePath).size;
@@ -16,7 +17,6 @@ const miloCSS = fs.readFileSync(miloCSSPath, 'utf8');
 
 logFileSize(miloCSSPath, 'original');
 
-// Formats and transforms Spectrum CSS Rgb and opacity properties for performance
 const transformRgbProperties = (rule) => {
   const properties = {};
 
@@ -24,7 +24,6 @@ const transformRgbProperties = (rule) => {
 
   rule.walkDecls((decl) => {
     const { prop, value } = decl;
-
     decl.value = replaceZeroPx(value);
 
     let baseName;
@@ -56,9 +55,7 @@ const extractAndTransform = (css, prefix) => {
     if (!customProperties[selector]) {
       customProperties[selector] = {};
     }
-
     transformRgbProperties(rule);
-
     rule.walkDecls((decl) => {
       let propName = decl.prop;
       if (propName.startsWith(prefix)) propName = `--spectrum${propName.slice(prefix.length)}`;
@@ -68,6 +65,7 @@ const extractAndTransform = (css, prefix) => {
   return customProperties;
 };
 
+// Merge additional properties into base properties
 const mergeProperties = (baseProps, additionalProps) => {
   Object.keys(additionalProps).forEach((prop) => {
     if (!baseProps[prop]) {
@@ -78,12 +76,17 @@ const mergeProperties = (baseProps, additionalProps) => {
 
 const spectrumProperties = extractAndTransform(spectrumCSS, '--spectrum');
 
+// Get Spectrum properties for a given selector
 const getSpectrumPropertiesForSelector = (selector) => {
   const properties = {};
-  if (selector === '.spectrum') {
+  if (selector === ':root') {
     mergeProperties(properties, spectrumProperties['.spectrum'] || {});
     mergeProperties(properties, spectrumProperties['.spectrum--light'] || {});
     mergeProperties(properties, spectrumProperties['.spectrum--medium'] || {});
+  } else if (selector === '.light') {
+    mergeProperties(properties, spectrumProperties['.spectrum--light'] || {});
+  } else if (selector === '.dark') {
+    mergeProperties(properties, spectrumProperties['.spectrum--dark'] || {});
   } else {
     mergeProperties(properties, spectrumProperties[selector] || {});
   }
@@ -102,7 +105,7 @@ const updateMiloCSS = (css) => {
         const { prop } = decl;
         if (prop.startsWith('--s2')) {
           const spectrumProp = `--spectrum${prop.slice(4)}`;
-          const spectrumSelector = selector.startsWith('.spectrum') ? selector : '.spectrum';
+          const spectrumSelector = selector === ':root' ? '.spectrum' : selector === '.light' ? '.spectrum--light' : selector === '.dark' ? '.spectrum--dark' : null;
           const spectrumProps = getSpectrumPropertiesForSelector(spectrumSelector);
           if (spectrumProps[spectrumProp] !== undefined) {
             decl.value = spectrumProps[spectrumProp];
